Cleaned up execution order
The run-game function was getting pretty messier.  Moved a lot of logic
in the run function to it's own functions. When this is all done pretty
much the only calls in the run function should be to initiate each stage
of the game engine actions (ie get input, validate, update)